+++ app/src/main/java/com/ugps/alcoolougasolina/interfaces/Callback.java
@@ -1,0 +1,7 @@
+package com.ugps.alcoolougasolina.interfaces;
+
+public interface Callback<T> {
+    void onResult(T data);
+
+    void onError(String error);
+}
