+++ app/src/main/java/com/ugps/alcoolougasolina/interfaces/ErrorListener.java
@@ -1,4 +1,4 @@
-package com.ugps.alcoolougasolina;
+package com.ugps.alcoolougasolina.interfaces;
 
 public interface ErrorListener {
     void onError(String error);
