+++ app/src/main/java/com/ugps/alcoolougasolina/models/OptionModel.java
@@ -1,4 +1,4 @@
-package com.ugps.alcoolougasolina;
+package com.ugps.alcoolougasolina.models;
 
 import androidx.annotation.NonNull;
