+++ app/src/main/java/com/ugps/alcoolougasolina/utils/CookieJar.java
@@ -1,4 +1,4 @@
-package com.ugps.alcoolougasolina;
+package com.ugps.alcoolougasolina.utils;
 
 import java.util.ArrayList;
 import java.util.List;
